moved transformations into geometry
